storage-initializer: Dynamic mount path
Right now, the StorageInitializer places PVCs under a predefined
path (`/mnt/pvc`). However, we want to place PVCs under the same path as
they were in the snapshotted Notebook, to ensure reproducibility. For
more information, see:
https://github.com/arrikto/dev/issues/815

To work around this limitation, we extend the mutating webhook in the
following way:
1. If the Pod has the `PVC_MOUNT_POINT` env var, we change the default
   mount point from `/mnt/pvc` to the env var's value.
2. Once we have done this operation, we skip adding the initContainer
   for copying data to a shared volume.

This is a hacky workaround, as it doesn't work with multiple PVCs.
Ideally, we would like something more generic, that would allow us to
work with any PVC.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Yannis Zarkadas <yanniszark@arrikto.com>